component moved into the directory
